moving rao parameters upload to file exporter
