+++ csa-runner-app/src/main/java/com/farao_community/farao/swe_csa/app/FileExporter.java
@@ -6,9 +6,14 @@ import com.powsybl.commons.datasource.MemDataSource;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.openrao.data.crac.api.Crac;
 import com.powsybl.openrao.data.raoresult.api.RaoResult;
+import com.powsybl.openrao.raoapi.json.JsonRaoParameters;
+import com.powsybl.openrao.raoapi.parameters.RaoParameters;
 import org.springframework.stereotype.Service;
 
 import java.io.*;
+import java.time.Instant;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
 import java.util.Properties;
 
 @Service
@@ -18,6 +23,8 @@ public class FileExporter {
     private static final String IIDM_EXTENSION = "xiidm";
 
     private final S3ArtifactsAdapter s3ArtifactsAdapter;
+
+    private static final DateTimeFormatter HOURLY_NAME_FORMATTER = DateTimeFormatter.ofPattern("yyyyMMdd'_'HHmm").withZone(ZoneId.of("UTC"));
 
     public FileExporter(S3ArtifactsAdapter s3ArtifactsAdapter) {
         this.s3ArtifactsAdapter = s3ArtifactsAdapter;
@@ -40,6 +47,15 @@ public class FileExporter {
         propertiesAmperes.setProperty("rao-result.export.json.flows-in-amperes", "true");
         raoResult.write("JSON", crac, propertiesAmperes, outputStreamRaoResult);
         s3ArtifactsAdapter.uploadFile(destinationPath, new ByteArrayInputStream(outputStreamRaoResult.toByteArray()));
+    }
+
+    public String uploadRaoParameters(Instant utcInstant, RaoParameters raoParameters) {
+        String raoParametersFilePath = String.format("configurations/rao-parameters-%s", HOURLY_NAME_FORMATTER.format(utcInstant).concat(".json"));
+        ByteArrayOutputStream baos = new ByteArrayOutputStream();
+        JsonRaoParameters.write(raoParameters, baos);
+        ByteArrayInputStream bais = new ByteArrayInputStream(baos.toByteArray());
+        s3ArtifactsAdapter.uploadFile(raoParametersFilePath, bais);
+        return s3ArtifactsAdapter.generatePreSignedUrl(raoParametersFilePath);
     }
 
 }
